Дорабатывает коннект пропсов
