+++ src/components/app/app.jsx
@@ -1,7 +1,5 @@
 import React, {PureComponent} from 'react';
 import {Switch, Route, BrowserRouter} from 'react-router-dom';
-import {connect} from 'react-redux';
-import PropTypes from 'prop-types';
 
 import Main from '../main/main.jsx';
 import MoviePage from '../movie-page/movie-page.jsx';
@@ -27,15 +25,8 @@ class App extends PureComponent {
   }
 
   _renderMain() {
-    const {film, movies} = this.props;
-    const {title, genre, year} = film;
-
     return (
       <Main
-        movieTitle={title}
-        movieGenre={genre}
-        movieYear={year}
-        movies={movies}
         onCardTitleClick={this._handleMovieCardClick}
         onCardClick={this._handleMovieCardClick}
       />
@@ -43,12 +34,8 @@ class App extends PureComponent {
   }
 
   _renderMoviePage() {
-    const {movies, film} = this.props;
-
     return (
       <MoviePageWrapped
-        film={film}
-        movies={movies}
         onCardClick={this._handleMovieCardClick}
         onCardTitleClick={this._handleMovieCardClick}
       />
@@ -81,33 +68,4 @@ class App extends PureComponent {
   }
 }
 
-App.propTypes = {
-  movies: PropTypes.arrayOf(
-      PropTypes.shape({
-        id: PropTypes.number.isRequired,
-        title: PropTypes.string.isRequired,
-        image: PropTypes.string.isRequired,
-      }).isRequired
-  ).isRequired,
-  film: PropTypes.shape({
-    title: PropTypes.string.isRequired,
-    genre: PropTypes.string.isRequired,
-    runTime: PropTypes.string.isRequired,
-    year: PropTypes.number.isRequired,
-    backgroundPoster: PropTypes.string.isRequired,
-    filmPoster: PropTypes.string.isRequired,
-    rating: PropTypes.number.isRequired,
-    ratingCount: PropTypes.number.isRequired,
-    description: PropTypes.string.isRequired,
-    director: PropTypes.string.isRequired,
-    starring: PropTypes.string.isRequired,
-  }).isRequired,
-};
-
-const mapStateToProps = (state) => ({
-  film: state.film,
-  movies: state.movies
-});
-
-export {App};
-export default connect(mapStateToProps)(App);
+export default App;
