+++ src/components/movie-page/movie-page.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import MoviePage from './movie-page.jsx';
+import {MoviePage} from './movie-page.jsx';
 import {MOVIES} from '../../common/consts';
 
 const film = {
@@ -18,23 +18,6 @@ const film = {
   starring: `Bill Murray, Edward Norton, Jude Law, Willem Dafoe and other`
 };
 
-const reviews = [
-  {
-    id: 1,
-    author: `Kate Muir`,
-    date: `December 24, 2016`,
-    rating: 8.9,
-    text: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years`,
-  },
-  {
-    id: 2,
-    author: `Bill Goodykoontz`,
-    date: `November 18, 2015`,
-    rating: 8.0,
-    text: `Anderson's films are too precious for some, but for those of us willing to lose ourselves in them, they're a delight. "The Grand Budapest Hotel" is no different, except that he has added a hint of gravitas to the mix, improving the recipe.`,
-  },
-];
-
 const activeTab = `Overview`;
 
 it(`Should MovieCard render correctly`, () => {
@@ -42,7 +25,6 @@ it(`Should MovieCard render correctly`, () => {
     .create(<MoviePage
       film={film}
       movies={MOVIES}
-      reviews={reviews}
       activeTab={activeTab}
       onCardClick={() => {}}
       onCardTitleClick={() => {}}
